arabic2vinculum: Make invalid states impossible
First of all, the return type of the function is Result, which is not
needed because we can *always* generate a roman number from any u64.

Second, the input is allowed to be zero and since this is not
representable, the implementation so far has returned an empty string.

Given that the return value is Result and the input can be zero, one
could possibly assume that whenever we pass a zero we get an Err, which
is not the case.

All this is really not necessary since we now use the NonZeroU64 type
which removes possible failures by design in our function.

Signed-off-by: aszlig <aszlig@nix.build>